requirements.txt
flask==3.0.2
python-dotenv==1.0.0
pydantic==2.6.1
langchain==0.1.9
langchain-google-genai==0.0.11
langchain-community==0.0.21
gunicorn==21.2.0